+++ packages/typespec-client-generator-core/src/types.ts
@@ -51,9 +51,11 @@ import {
   getAccessOverride,
   getAlternateType,
   getClientNamespace,
+  getExplicitClientApiVersions,
   getOverriddenClientMethod,
   getUsageOverride,
   listClients,
+  listOperationGroups,
   listOperationsInOperationGroup,
   shouldFlattenProperty,
   shouldGenerateConvenient,
@@ -104,7 +106,7 @@ import {
   isNeverOrVoidType,
   isOnClient,
   listAllUserDefinedNamespaces,
-  listRawSubClients,
+  resolveConflictGeneratedName,
   twoParamsEquivalent,
   updateWithApiVersionInformation,
 } from "./internal-utils.js";
@@ -497,7 +499,7 @@ export function getSdkUnionWithDiagnostics(
   type: Union,
   operation?: Operation,
 ): [SdkType, readonly Diagnostic[]] {
-  let retval: SdkType | undefined = context.__referencedTypeCache?.get(type);
+  let retval: SdkType | undefined = context.__referencedTypeCache.get(type);
   const diagnostics = createDiagnosticCollector();
 
   if (!retval) {
@@ -807,11 +809,10 @@ export function getSdkModelWithDiagnostics(
   operation?: Operation,
 ): [SdkModelType, readonly Diagnostic[]] {
   const diagnostics = createDiagnosticCollector();
-  let sdkType = context.__referencedTypeCache?.get(type) as SdkModelType | undefined;
+  let sdkType = context.__referencedTypeCache.get(type) as SdkModelType | undefined;
 
   if (!sdkType) {
     const name = getLibraryName(context, type) || getGeneratedName(context, type, operation);
-    const usage = isErrorModel(context.program, type) ? UsageFlags.Error : UsageFlags.None; // eslint-disable-line @typescript-eslint/no-deprecated
     sdkType = {
       ...diagnostics.pipe(getSdkTypeBaseHelper(context, type, "model")),
       name: name,
@@ -822,7 +823,7 @@ export function getSdkModelWithDiagnostics(
       properties: [],
       additionalProperties: undefined, // going to set additional properties in the next few lines when we look at base model
       access: "public",
-      usage,
+      usage: UsageFlags.None,
       crossLanguageDefinitionId: getCrossLanguageDefinitionId(context, type, operation),
       apiVersions: getAvailableApiVersions(context, type, type.namespace),
       serializationOptions: {},
@@ -844,7 +845,7 @@ export function getSdkModelWithDiagnostics(
     // propreties should be generated first since base model'sdiscriminator handling is depend on derived model's properties
     diagnostics.pipe(addPropertiesToModelType(context, type, sdkType, operation));
     if (type.baseModel) {
-      sdkType.baseModel = context.__referencedTypeCache?.get(type.baseModel) as
+      sdkType.baseModel = context.__referencedTypeCache.get(type.baseModel) as
         | SdkModelType
         | undefined;
       if (sdkType.baseModel === undefined) {
@@ -941,7 +942,7 @@ function getSdkEnumWithDiagnostics(
   operation?: Operation,
 ): [SdkEnumType, readonly Diagnostic[]] {
   const diagnostics = createDiagnosticCollector();
-  let sdkType = context.__referencedTypeCache?.get(type) as SdkEnumType | undefined;
+  let sdkType = context.__referencedTypeCache.get(type) as SdkEnumType | undefined;
   if (!sdkType) {
     sdkType = {
       ...diagnostics.pipe(getSdkTypeBaseHelper(context, type, "enum")),
@@ -1339,32 +1340,6 @@ function updateMultiPartInfo(
     if (httpPart?.options?.name) {
       base.serializedName = httpPart?.options?.name; // eslint-disable-line @typescript-eslint/no-deprecated
       base.serializationOptions.multipart.name = httpPart?.options?.name;
-    }
-  } else {
-    // common body
-    const httpOperation = getHttpOperationWithCache(context, operation);
-    const operationIsMultipart = Boolean(
-      httpOperation && httpOperation.parameters.body?.contentTypes.includes("multipart/form-data"),
-    );
-    if (operationIsMultipart) {
-      const isBytesInput =
-        base.type.kind === "bytes" ||
-        (base.type.kind === "array" && base.type.valueType.kind === "bytes");
-      // Currently we only recognize bytes and list of bytes as potential file inputs
-      if (isBytesInput && getEncode(context.program, type)) {
-        diagnostics.add(
-          createDiagnostic({
-            code: "encoding-multipart-bytes",
-            target: type,
-          }),
-        );
-      }
-      base.serializationOptions.multipart = {
-        isFilePart: isBytesInput,
-        isMulti: base.type.kind === "array",
-        defaultContentTypes: [],
-        name: base.name,
-      };
     }
   }
   if (base.serializationOptions.multipart !== undefined) {
@@ -1811,13 +1786,13 @@ function updateTypesFromOperation(
 function updateAccessOverride(context: TCGCContext): [void, readonly Diagnostic[]] {
   const diagnostics = createDiagnosticCollector();
   // set access for all orphan model without override
-  for (const sdkType of context.__referencedTypeCache?.values() ?? []) {
+  for (const sdkType of context.__referencedTypeCache.values()) {
     const accessOverride = getAccessOverride(context, sdkType.__raw as any);
     if (!sdkType.__accessSet && accessOverride === undefined) {
       diagnostics.pipe(updateUsageOrAccess(context, "public", sdkType));
     }
   }
-  for (const sdkType of context.__referencedTypeCache?.values() ?? []) {
+  for (const sdkType of context.__referencedTypeCache.values()) {
     const accessOverride = getAccessOverride(context, sdkType.__raw as any);
     if (accessOverride) {
       diagnostics.pipe(updateUsageOrAccess(context, accessOverride, sdkType, { isOverride: true }));
@@ -1828,7 +1803,7 @@ function updateAccessOverride(context: TCGCContext): [void, readonly Diagnostic[
 
 function updateUsageOverride(context: TCGCContext): [void, readonly Diagnostic[]] {
   const diagnostics = createDiagnosticCollector();
-  for (const sdkType of context.__referencedTypeCache?.values() ?? []) {
+  for (const sdkType of context.__referencedTypeCache.values()) {
     const usageOverride = getUsageOverride(context, sdkType.__raw as any);
     if (usageOverride) {
       diagnostics.pipe(updateUsageOrAccess(context, usageOverride, sdkType, { isOverride: true }));
@@ -1838,7 +1813,7 @@ function updateUsageOverride(context: TCGCContext): [void, readonly Diagnostic[]
 }
 
 function updateSpreadModelUsageAndAccess(context: TCGCContext): void {
-  for (const [_, sdkType] of context.__referencedTypeCache?.entries() ?? []) {
+  for (const [_, sdkType] of context.__referencedTypeCache.entries()) {
     if (
       sdkType.kind === "model" &&
       (sdkType.usage & UsageFlags.Spread) > 0 &&
@@ -1876,7 +1851,7 @@ function filterOutTypes(
   filter: number,
 ): (SdkModelType | SdkEnumType | SdkUnionType | SdkNullableType)[] {
   const result = new Array<SdkModelType | SdkEnumType | SdkUnionType | SdkNullableType>();
-  for (const sdkType of context.__referencedTypeCache?.values() ?? []) {
+  for (const sdkType of context.__referencedTypeCache.values()) {
     // filter models with unexpected usage
     if ((sdkType.usage & filter) === 0) {
       continue;
@@ -1937,7 +1912,7 @@ export function handleAllTypes(context: TCGCContext): [void, readonly Diagnostic
       // operations on a client
       diagnostics.pipe(updateTypesFromOperation(context, operation));
     }
-    for (const sc of listRawSubClients(context, client)) {
+    for (const sc of listOperationGroups(context, client, true)) {
       for (const operation of listOperationsInOperationGroup(context, sc)) {
         // operations on operation groups
         diagnostics.pipe(updateTypesFromOperation(context, operation));
@@ -1955,9 +1930,16 @@ export function handleAllTypes(context: TCGCContext): [void, readonly Diagnostic
     const [_, versionMap] = getVersions(context.program, client.service);
     if (versionMap && versionMap.getVersions()[0]) {
       // create sdk enum for versions enum
-      const sdkVersionsEnum = diagnostics.pipe(
-        getSdkEnumWithDiagnostics(context, versionMap.getVersions()[0].enumMember.enum),
-      );
+      let sdkVersionsEnum: SdkEnumType;
+      const explicitApiVersions = getExplicitClientApiVersions(context, client.service);
+      if (explicitApiVersions) {
+        // add additional api versions to the enum
+        sdkVersionsEnum = diagnostics.pipe(getSdkEnumWithDiagnostics(context, explicitApiVersions));
+      } else {
+        sdkVersionsEnum = diagnostics.pipe(
+          getSdkEnumWithDiagnostics(context, versionMap.getVersions()[0].enumMember.enum),
+        );
+      }
       filterApiVersionsInEnum(context, client, sdkVersionsEnum);
       diagnostics.pipe(updateUsageOrAccess(context, UsageFlags.ApiVersionEnum, sdkVersionsEnum));
     }
@@ -1989,6 +1971,9 @@ export function handleAllTypes(context: TCGCContext): [void, readonly Diagnostic
   diagnostics.pipe(updateUsageOverride(context));
   // update spread model
   updateSpreadModelUsageAndAccess(context);
+
+  // update generated name
+  resolveConflictGeneratedName(context);
 
   return diagnostics.wrap(undefined);
 }
@@ -2070,7 +2055,7 @@ function setSerializationOptions(
   type: SdkModelType | SdkBodyModelPropertyType,
   contentTypes: string[],
 ) {
-  for (const contentType of contentTypes ?? []) {
+  for (const contentType of contentTypes) {
     if (isMediaTypeJson(contentType) && !type.serializationOptions.json) {
       updateJsonSerializationOptions(context, type);
     }