+++ packages/typespec-client-generator-core/test/test-host.ts
@@ -11,11 +11,12 @@ import {
 import { HttpTestLibrary } from "@typespec/http/testing";
 import { RestTestLibrary } from "@typespec/rest/testing";
 import { VersioningTestLibrary } from "@typespec/versioning/testing";
-import { CreateSdkContextOptions, SdkEmitterOptions, createSdkContext } from "../src/context.js";
+import { CreateSdkContextOptions, createSdkContext } from "../src/context.js";
 import { SdkContext, SdkHttpOperation, SdkServiceOperation } from "../src/interfaces.js";
+import { BrandedSdkEmitterOptionsInterface } from "../src/internal-utils.js";
 import { SdkTestLibrary } from "../src/testing/index.js";
 
-export interface CreateSdkTestRunnerOptions extends SdkEmitterOptions {
+export interface CreateSdkTestRunnerOptions extends BrandedSdkEmitterOptionsInterface {
   emitterName?: string;
   librariesToAdd?: TypeSpecTestLibrary[];
   autoImports?: string[];
@@ -203,7 +204,7 @@ export async function createSdkTestRunner(
       .filter((x) => x !== StandardTestLibrary)
       .map((x) => x.name)
       .map((x) => `import "${x}";`),
-    ...(autoUsings ?? []).map((x) => `using ${x};`),
+    ...autoUsings.map((x) => `using ${x};`),
   ].join("\n");
 
   const clientAutoCode = [
@@ -212,7 +213,7 @@ export async function createSdkTestRunner(
       .map((x) => x.name)
       .map((x) => `import "${x}";`),
     `import "./main.tsp";`,
-    ...(autoUsings ?? []).map((x) => `using ${x};`),
+    ...autoUsings.map((x) => `using ${x};`),
   ].join("\n");
 
   // compile with client.tsp